fixed possible leak, x64 still not working

+++ SpoutControls4vvvv/SpoutControls4vvvv.h
@@ -23,13 +23,13 @@ typedef struct MYSTRINGS
 
 } MYSTRINGS;
 
-extern "C" { __declspec(dllexport) void InitializeControls(char *sendername, int *numControls, char** names, int *types, double* floats, double* toggles, double* press, char** text); }
+extern "C" { __declspec(dllexport) void InitializeControls(char *sendername, int *numControls, char** names, int *types, float* floats, float* toggles, float* press, char** text); }
 //extern "C" { __declspec(dllexport) void InitializeControls(MYCONTROLS* myC); }
-extern "C" { __declspec(dllexport) bool UpdateControls(const char** text, double *floats, double *toggles, double *press, int *numControls); }
-extern "C" { __declspec(dllexport) void  UpdateFloatControls(int *numControls, double * floats); }
+extern "C" { __declspec(dllexport) bool UpdateControls(const char** text, float *floats, float *toggles, float *press, int *numControls); }
+extern "C" { __declspec(dllexport) void  UpdateFloatControls(int *numControls, float * floats); }
 extern "C" { __declspec(dllexport) void CloseControls(); }
 //
-extern "C" __declspec(dllexport) int ReleaseMemory(double *pArray)
+extern "C" __declspec(dllexport) int ReleaseMemory(float *pArray)
 {
 	delete[] pArray;
 	//delete[] Usize;
